12/04 image 넣어서 character 구현 및 Enemy(Subject) 구현

+++ Enemy.py
@@ -11,7 +11,8 @@ class Enemy:
         self.appearance = 'circle'
         self.state = 'alive'
         self.position = np.array([spawn_position[0] - 25, spawn_position[1] - 25, spawn_position[0] + 25, spawn_position[1] + 25])
-        self.center = np.array([(self.position[0] + self.position[2]) / 2, (self.position[1] + self.position[3]) / 2])
+        #self.center = np.array([(self.position[0] + self.position[2]) / 2, (self.position[1] + self.position[3]) / 2])
+        self.center = np.array([int((self.position[0] + self.position[2]) / 2), int((self.position[1] + self.position[3]) / 2)])
         self.outline = "#00FF00"
 
         self.die_flag=die_flag
@@ -37,7 +38,8 @@ class Enemy:
         self.appearance = 'circle'
         self.state = 'alive'
         self.position = np.array([spawn_position[0] - 25, spawn_position[1] - 25, spawn_position[0] + 25, spawn_position[1] + 25])
-        self.center = np.array([(self.position[0] + self.position[2]) / 2, (self.position[1] + self.position[3]) / 2])
+        #self.center = np.array([(self.position[0] + self.position[2]) / 2, (self.position[1] + self.position[3]) / 2])
+        self.center = np.array([int((self.position[0] + self.position[2]) / 2), int((self.position[1] + self.position[3]) / 2)])
         self.outline = "#00FF00"
 
         self.die_flag=die_flag
